Rename dir name from lib to src
