+++ src/routes/terms.js
